+++ app/src/main/java/edu/wmdd/gover/InspectionDetailActivity.java
@@ -64,7 +64,7 @@ public class InspectionDetailActivity extends AppCompatActivity {
 //    private SignaturePad mSignaturePad;
 //    private Button mClearButton;
 //    private Button mSaveButton;
-    Bitmap SignatureBitmap;
+
     String encodedSignature;
 
     //================
@@ -111,62 +111,7 @@ public class InspectionDetailActivity extends AppCompatActivity {
             }
         });
 
-        //signature
-        btnAddSignature = findViewById(R.id.addSignature);
-        btnAddSignature.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                final Dialog dialog = new Dialog(context);
-                dialog.setContentView(R.layout.signature_layout);
 
-                Button mClearButton = (Button) dialog.findViewById(R.id.clear_button);
-                Button mSaveButton = (Button) dialog.findViewById(R.id.save_button);
-
-                SignaturePad mSignaturePad = (SignaturePad) dialog.findViewById(R.id.signature_pad);
-                mSignaturePad.setOnSignedListener(new SignaturePad.OnSignedListener() {
-
-                    @Override
-                    public void onStartSigning() {
-                        //Event triggered when the pad is touched
-
-
-                    }
-
-                    @Override
-                    public void onSigned() {
-                        //Event triggered when the pad is signed
-                        mSaveButton.setEnabled(true);
-                        mClearButton.setEnabled(true);
-                    }
-
-                    @Override
-                    public void onClear() {
-                        //Event triggered when the pad is cleared
-                        mSaveButton.setEnabled(false);
-                        mClearButton.setEnabled(false);
-                    }
-                });
-
-                mSaveButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        SignatureBitmap = mSignaturePad.getSignatureBitmap();
-                        bitmapToBase64();
-//                        base64ToImage();
-                    }
-                });
-                mClearButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        mSignaturePad.clear();
-                    }
-                });
-
-                dialog.show();
-            }
-        });
-
-        //=====
 
         btReport = (Button) findViewById(R.id.btReport);
         btReport.setOnClickListener(new View.OnClickListener() {
@@ -180,13 +125,13 @@ public class InspectionDetailActivity extends AppCompatActivity {
 
         manageScreenState();
     }
-    public void bitmapToBase64() {
-        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
-        SignatureBitmap.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
-        byte[] byteArray = byteArrayOutputStream .toByteArray();
-        encodedSignature = Base64.encodeToString(byteArray, Base64.DEFAULT);
-        Log.d("base64", Base64.encodeToString(byteArray, Base64.DEFAULT));
-    }
+//    public void bitmapToBase64() {
+//        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+//        SignatureBitmap.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
+//        byte[] byteArray = byteArrayOutputStream .toByteArray();
+//        encodedSignature = Base64.encodeToString(byteArray, Base64.DEFAULT);
+//        Log.d("base64", Base64.encodeToString(byteArray, Base64.DEFAULT));
+//    }
 
     //check if an inspection was sent (from Inspection List) or a property (from Create Inspection property list)
     private void manageScreenState() {
