+++ app/src/main/res/layout/activity_inspection_detail.xml
@@ -61,15 +61,5 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent" />
 
-    <Button
-        android:id="@+id/addSignature"
-        android:layout_width="144dp"
-        android:layout_height="51dp"
-        android:layout_marginStart="44dp"
-        android:layout_marginBottom="16dp"
-        android:text="addSignature"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
